Added React Frontend

+++ Backend/controllers/goalControllers.js
@@ -4,9 +4,10 @@ const asyncHandler = require('express-async-handler');
 // @route GET / api/goals
 // @access private
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 const getGoals = asyncHandler( async (req,res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     res.status(200).json(goals)
 })
 // @desc   Set Goals
@@ -18,9 +19,9 @@ const setGoals = asyncHandler( async (req,res) => {
         res.status(400)
         throw new Error('Please add a text field')
     }
-
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id,
     })
 
     res.status(200).json(goal)
@@ -37,7 +38,17 @@ const updateGoals = asyncHandler( async (req,res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
-
+    const user = await User.findById(req.user.id)
+    //check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    // make sure the logedin user matches the goal user
+    if(goal.user.toString() !=  user.id) {
+        res.status(401)
+        throw new Error('User not Authorized')
+    }
     const updateGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new:true,
     })
@@ -56,12 +67,22 @@ const deleteGoals = asyncHandler(async (req, res) => {
         res.status(400);
         throw new Error('Goal not found');
     }
+    const user = await User.findById(req.user.id)
+    //check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    // make sure the logedin user matches the goal user
+    if(goal.user.toString() !=  user.id) {
+        res.status(401)
+        throw new Error('User not Authorized')
+    }
 
     const deletedGoal = await Goal.findByIdAndDelete(req.params.id);
 
     res.status(200).json({ message: `Deleted Goal: ${deletedGoal}` });
 });
-
 
 module.exports = {
     getGoals, setGoals, updateGoals, deleteGoals
